LayoutTests/media/media-initialTime-expected.txt

Test the, so far unused, 'initialTime' attribute.

EVENT(loadeddata)
EXPECTED (video.initialTime == '0') OK
RUN(video.initialTime = 10)
EXPECTED (video.initialTime == '0') OK
END OF TEST
